# packages/docs/openapi/data-engine.yaml
openapi: 3.0.0
info:
    title: Data Engine API
    description: |
        API for managing structured data resources including domains, URLs, calls, tags, documents, tickets, and chats.

        This API provides comprehensive CRUD operations for managing data across multiple domains with support for filtering, pagination, and advanced search capabilities.
    version: 1.0.0
    contact:
        name: AnswerAgentAI Support
        url: https://theanswer.ai
        email: support@theanswer.ai
servers:
    - url: https://prod.studio.theanswer.ai/api/v1/data-engine
      description: Production server
    - url: http://localhost:4000/api/v1/data-engine
      description: Local development server
...
paths:
...
    /calls:
        post:
            summary: Create a new call
            description: Create a new call log record with transcript and metadata
            operationId: createCall
            tags:
                - calls
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CallInput'
            responses:
                '201':
                    description: Call created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Call'
        get:
            summary: List all calls
            description: Retrieve calls with filtering and pagination
            operationId: listCalls
            tags:
                - calls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - $ref: '#/components/parameters/SearchTermParam'
                - name: employeeName
                  in: query
                  schema:
                      type: string
                - name: callerName
                  in: query
                  schema:
                      type: string
                - name: callType
                  in: query
                  schema:
                      type: string
                - name: sentimentMin
                  in: query
                  schema:
                      type: number
                      format: float
                - name: sentimentMax
                  in: query
                  schema:
                      type: number
                      format: float
                - name: dateFrom
                  in: query
                  schema:
                      type: string
                      format: date-time
                - name: dateTo
                  in: query
                  schema:
                      type: string
                      format: date-time
                - name: hasTranscript
                  in: query
                  schema:
                      type: boolean
                - name: hasCoaching
                  in: query
                  schema:
                      type: boolean
                - name: includeTags
                  in: query
                  schema:
                      type: string
                - name: excludeTags
                  in: query
                  schema:
                      type: string
                - name: customFilters
                  in: query
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CallListResponse'
    /calls/{id}:
        get:
            summary: Get call by ID
            operationId: getCallById
            tags:
                - calls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Call'
        put:
            summary: Update call
            operationId: updateCall
            tags:
                - calls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CallUpdate'
            responses:
                '200':
                    description: Call updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Call'
        delete:
            summary: Delete call
            operationId: deleteCall
            tags:
                - calls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Call deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
    # ==================== TAGS ====================
...
    /chats:
        post:
            summary: Create a new chat
            description: Create a new chat log record
            operationId: createChat
            tags:
                - chats
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChatInput'
            responses:
                '201':
                    description: Chat created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Chat'
        get:
            summary: List all chats
            operationId: listChats
            tags:
                - chats
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - name: chatbotName
                  in: query
                  schema:
                      type: string
                - name: aiModel
                  in: query
                  schema:
                      type: string
                - name: tags
                  in: query
                  description: Array of tag strings (JSON string)
                  schema:
                      type: string
                - name: sentimentMin
                  in: query
                  schema:
                      type: number
                      format: float
                - name: sentimentMax
                  in: query
                  schema:
                      type: number
                      format: float
                - name: resolutionStatus
                  in: query
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ChatListResponse'
    /chats/{id}:
        get:
            summary: Get chat by ID
            operationId: getChatById
            tags:
                - chats
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Chat'
        put:
            summary: Update chat
            operationId: updateChat
            tags:
                - chats
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChatUpdate'
            responses:
                '200':
                    description: Chat updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Chat'
        delete:
            summary: Delete chat
            operationId: deleteChat
            tags:
                - chats
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Chat deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: API Key
            description: |
                API key authentication. Include your API key in the Authorization header:
                `Authorization: Bearer YOUR_API_KEY`
    parameters:
        PageParam:
            name: page
            in: query
            description: Page number (0-indexed)
            schema:
                type: integer
                default: 0
                minimum: 0
        PageSizeParam:
            name: pageSize
            in: query
            description: Items per page
            schema:
                type: integer
                default: 10
                minimum: 1
                maximum: 100
        SearchTermParam:
            name: searchTerm
            in: query
            description: Search term for filtering results
            schema:
                type: string
        IdParam:
            name: id
            in: path
            required: true
            description: Resource ID (UUID)
            schema:
                type: string
                format: uuid
    schemas:
...
        Metadata:
            type: object
            description: Metadata object for tracking source system and user information
            properties:
                source_system:
                    type: string
                    example: theanswer
                source_organization_id:
                    type: string
                    format: uuid
                source_user_id:
                    type: string
                    format: uuid
                created_by:
                    type: string
                    format: email
                last_updated_by:
                    type: string
                    format: email
                last_updated_from:
                    type: string
                    example: theanswer
        DeleteResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: Resource deleted successfully
        PaginationInfo:
            type: object
            properties:
                total:
                    type: integer
                    example: 100
                page:
                    type: integer
                    example: 0
                pageSize:
                    type: integer
                    example: 10
        # Domain Schemas
...
        CallInput:
            type: object
            required:
                - transcript
            properties:
                transcript:
                    type: string
                    example: Full call transcript text...
                transcript_json:
                    type: object
                    additionalProperties: true
                    example:
                        speakers: [Agent, Caller]
                        segments:
                            - speaker: Agent
                              text: Hello, how can I help?
                              timestamp: 0.0
                recording_url:
                    type: string
                    format: uri
                    example: https://example.com/recording.mp3
                duration:
                    type: integer
                    example: 300
                call_datetime:
                    type: string
                    format: date-time
                    example: '2025-01-20T10:00:00Z'
                sentiment_score:
                    type: number
                    format: float
                    example: 7.5
                summary:
                    type: string
                    example: Call summary text
                custom_data:
                    type: object
                    additionalProperties: true
                    example:
                        employee_name: John Doe
                        caller_name: Jane Smith
                        call_type: inbound
                ai_analysis:
                    type: object
                    additionalProperties: true
                ai_coaching:
                    type: object
                    additionalProperties: true
                    example:
                        suggestions: [Be more empathetic]
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Call:
            allOf:
                - $ref: '#/components/schemas/CallInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        CallUpdate:
            type: object
            properties:
                sentiment_score:
                    type: number
                    format: float
                summary:
                    type: string
                custom_data:
                    type: object
                    additionalProperties: true
                metadata:
                    $ref: '#/components/schemas/Metadata'
        CallListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      calls:
                          type: array
                          items:
                              $ref: '#/components/schemas/Call'
        # Tag Schemas
...
        ChatInput:
            type: object
            required:
                - chatbot_name
                - user_message
                - ai_response
            properties:
                chatbot_name:
                    type: string
                    example: Support Bot
                ai_model:
                    type: string
                    example: gpt-4
                conversation_id:
                    type: string
                    example: conv-123
                user_message:
                    type: string
                    example: Hello
                ai_response:
                    type: string
                    example: Hi! How can I help?
                sentiment_score:
                    type: number
                    format: float
                    example: 7.5
                resolution_status:
                    type: string
                    example: resolved
                tags_array:
                    type: array
                    items:
                        type: string
                    example: [support, greeting]
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Chat:
            allOf:
                - $ref: '#/components/schemas/ChatInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        ChatUpdate:
            type: object
            properties:
                sentiment_score:
                    type: number
                    format: float
                resolution_status:
                    type: string
                tags_array:
                    type: array
                    items:
                        type: string
                metadata:
                    $ref: '#/components/schemas/Metadata'
        ChatListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      chats:
                          type: array
                          items:
                              $ref: '#/components/schemas/Chat'
...